Break identify route from server.js
